gcp/requirements.txt
# GCP Cloud Function dependencies
functions-framework==3.5.0
Pillow==10.2.0
numpy==1.26.4
ai-edge-litert
